More uniformity in do blocks.
Also unless can be used outside of do.

Fixes some things, breaks a few other things...

In particular I can't figure out how to get try/catch/finally to parse
the finally as not-a-function-apply without some nonsense.

Introducing _primary_expression also was necessary to pass these...

So this is almost definitely not completely right, but it's a step
forward overall I think.