Binders via the lean parser structure.
This.. breaks the Structure and Class parsing and I
can't figure out quite why. I guess adding indentation
support may fix it. What's wrong at the minute is that
an apply in a field value will get parsed as a new simple
binder, so possibly the issue is with apply's binding
strength as well, which I'll try and fix as things get
closer to the upstream parser.